# Optimize bundle-add