--- a/src/list.h
+++ b/src/list.h
@@ -51,4 +51,8 @@ struct list *list_clone(struct list *list);
 
 /* deep copy of a string list */
 struct list *list_deep_clone_strs(struct list *source);
+
+/* check list length */
+extern int list_longer_than(struct list *list, int count);
+
 #endif