--- a/test/functional/bundleadd/add-rc/web-dir/10/files/6c27df6efcd6fc401ff1bc67c970b83eef115f6473db4fb9d57e5de317eba96e/.gitignore
+++ b/test/functional/bundleadd/add-rc/web-dir/10/files/6c27df6efcd6fc401ff1bc67c970b83eef115f6473db4fb9d57e5de317eba96e/.gitignore
@@ -1,0 +1,4 @@
+# Ignore everything in this directory
+*
+# Except this file
+!.gitignore