--- a/test/functional/bundleadd/add-rc/web-dir/10/files/e6d85023c5e619eb43d5cfbfdbdec784afef5a82ffa54e8c93bda3e0883360a3
+++ b/test/functional/bundleadd/add-rc/web-dir/10/files/e6d85023c5e619eb43d5cfbfdbdec784afef5a82ffa54e8c93bda3e0883360a3
@@ -1,0 +1,1 @@
+test boot file