--- a/test/functional/bundleadd/fix-missing-file/lines-checked
+++ b/test/functional/bundleadd/fix-missing-file/lines-checked
@@ -1,4 +1,3 @@
-Downloading packs
 Starting download of remaining update content This may take a while
 File /foo was not in a pack
 Finishing download of update content