--- a/test/functional/bundleadd/skip-scripts/test.bats
+++ b/test/functional/bundleadd/skip-scripts/test.bats
@@ -10,6 +10,7 @@ setup() {
   sign_manifest_mom 10
   create_manifest_tar 10 os-core
   create_manifest_tar 10 test-bundle
+  create_fullfile_tar 10 $f_hash
   chown_root "$DIR/web-dir/10/staged/$f_hash"
   tar -C "$DIR/web-dir/10" -cf "$DIR/web-dir/10/pack-test-bundle-from-0.tar" staged/$f_hash
 }