--- a/test/functional/bundleadd/verify-fix-path/lines-checked
+++ b/test/functional/bundleadd/verify-fix-path/lines-checked
@@ -1,7 +1,8 @@
-Extracting test-bundle pack for version 10
 Starting download of remaining update content. This may take a while...
+File /usr/bin/foo was not in a pack
 Finishing download of update content...
 Installing bundle(s) files...
-REGEXP:Update target directory does not exist: .*/usr/bin. Trying to fix it
+REGEXP:Update target directory does not exist: .*/usr/bin. .*
 Path /usr/bin is missing on the file system ... fixing
+Path /usr/bin/foo is missing on the file system ... fixing
 Bundle(s) installation done.