* Don't install files that are already present
* Don't parse the whole Manifest-tree of all installed manifest unless and until error recovery is required

with these changes a swupd bundle-add git becomes a 416 msec thing of which +/- 200 msec is the sync() and 180 or so msec are the triggers


Raw elapsed time stats:
      0.76 ms: Add bundles and recurse
      3.28 ms: Consolidate files from bundles
      0.03 ms: Download packs
      0.05 ms: Download missing files
    236.81 ms: Installing bundle(s) files onto filesystem

CPU process time stats:
      0.75 ms: Add bundles and recurse
      3.27 ms: Consolidate files from bundles
      0.03 ms: Download packs
      0.05 ms: Download missing files
      0.35 ms: Installing bundle(s) files onto filesystem

Raw elapsed time stats:
      0.00 ms: Prepend bundles to list
    416.82 ms: Install bundles

CPU process time stats:
      0.00 ms: Prepend bundles to list
      4.99 ms: Install bundles