@matthewrsj 2018-06-11
Fixes #478 

I'll look into test failures today.