@matthewrsj 2018-06-11
Ah, many tests are failing due to output-matching for pack downloads. In many cases packs aren't being attempted because they provide less than 5 files.