@icarus-sparry 2018-06-11
Another argument for improving our test tooling.