@bryteise 2018-06-12
@matthewrsj One thing I'm worried about with this change is that the pack download and extraction might not being used in any tests. Do we have some tests where the 0-pack is used still?